Tests started

+++ employeePayroll/src/main/java/employeePayroll/PayScheme.java
@@ -2,5 +2,5 @@ package employeePayroll;
 
 public abstract class PayScheme {
 
-	abstract public float payEmployee(float hourRate, float hoursWorked);
+	abstract public float payEmployee(double hourRate, double hoursWorked);
 }
